Convert values to 64-bit types
Change ints and floats used in values handled by calculator
to long long ints and doubles which are 64-bit values on
target platforms.

+++ lib/InfixParser.c
@@ -296,7 +296,7 @@ memerr:
 
 static SilikoSyntaxTreeNode *GetExpr3lf(SilikoLexer *lexer)
 {
-	int value;
+	long long int value;
 
 	if(lexer->Token.Type == 'd')
 	{
